Added 'touch-like' command

+++ tests/test_fs.py
@@ -5,14 +5,19 @@ from nose.tools import ok_
 from nose.tools import raises
 import os
 from py_utilities.fs import mkdir_p
+from py_utilities.fs import touch
 import shutil
 import unittest
 
 
 class TestFs(unittest.TestCase):
 
+    def setUp(self):
+        self.cwd = os.path.dirname(os.path.realpath(__file__))
+
     def tearDown(self):
-        shutil.rmtree('/tmp/foo')
+        if os.path.exists('/tmp/foo'):
+            shutil.rmtree('/tmp/foo')
 
     def test_mkdir_p(self):
         mkdir_p('/tmp/foo/bar/rules')
@@ -31,5 +36,11 @@ class TestFs(unittest.TestCase):
         mkdir_p('/tmp/foo/bar/rules')
         mkdir_p('/tmp/foo/bar/rules')
 
+    def test_touch(self):
+        file = os.path.join(self.cwd, 'test_fs.py')
+        curr_access_time = os.path.getmtime(file)
+        touch(file)
+        new_access_time = os.path.getmtime(file)
+        ok_(new_access_time > curr_access_time)
 
 # vim: filetype=python
